Hotfix exercise 1

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -3,13 +3,13 @@
 /**
  * sum_them_all - return the sum of all its parameters
  * @n: number of parameters
- * Return: sum or 0 if n == 0
+ * Return: int
  */
 int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
 	unsigned int i;
-	long int sum = 0;
+	int sum = 0;
 
 	if (n == 0)
 		return (0);
